(QENG-4283) Support pre-generated Beaker hosts file
Prior to this commit, only the client test target was configurable when running
the acceptance tests. The master test target was hardcoded, and the vmpooler
hypervisor was always used by Beaker (by virtue of not specifying a hypervisor,
resulting the in the default vmpooler one).

In order to run in CI.next though, all test targets must be known ahead of time,
and the ABS hypervisor must be used by Beaker.

This commit allows for a pre-generated Beaker hosts file to be used, by
specifying it as the value of the BEAKER_HOSTS environment variable.

If BEAKER_HOSTS is not defined, the original behavior will be used.